Fix: notebook_login() does not update UI on Databricks (#1414)
* wip

* catch errors to display

* change output

* display error

* change error output

* refactor

* support multiple string objects

* fix strings and add typing

* black

* style

* Use capture_output in notebook_login

* Fix messages

* fix tests

---------
