+++ src/huggingface_hub/utils/_subprocess.py
@@ -16,7 +16,9 @@
 """Contains utilities to easily handle subprocesses in `huggingface_hub`."""
 import os
 import subprocess
+import sys
 from contextlib import contextmanager
+from io import StringIO
 from pathlib import Path
 from typing import IO, Generator, List, Optional, Tuple, Union
 
@@ -24,6 +26,26 @@ from .logging import get_logger
 
 
 logger = get_logger(__name__)
+
+
+@contextmanager
+def capture_output() -> Generator[StringIO, None, None]:
+    """Capture output that is printed to terminal.
+
+    Taken from https://stackoverflow.com/a/34738440
+
+    Example:
+    ```py
+    >>> with capture_output() as output:
+    ...     print("hello world")
+    >>> assert output.getvalue() == "hello world\n"
+    ```
+    """
+    output = StringIO()
+    previous_output = sys.stdout
+    sys.stdout = output
+    yield output
+    sys.stdout = previous_output
 
 
 def run_subprocess(
